requirements.txt
requests
datamule
PyQt6
dotenv
google-genai